requirements.txt
Django==1.11.15
djangorestframework==3.9.0
mod-wsgi==4.6.2
python-magic==0.4.15
pytz==2017.3
